requirements.txt
requests
pandas
pathlib
typing
python-dotenv